requirements.txt
streamlit
pillow
piexif
